Minimize code (refactor)
